refactor: build event
