+++ cmd/meroxa/builder/builder.go
@@ -22,7 +22,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"runtime"
 	"strings"
 	"time"
 
@@ -332,61 +331,6 @@ func buildCommandWithDocs(cmd *cobra.Command, c Command) {
 	cmd.Example = docs.Example
 }
 
-func getCLIUserInfo() (actor, actorUUID string, err error) {
-	// Require login
-	_, _, err = global.RequireLogin()
-
-	/*
-		 	We don't report client issues to the customer as it'll likely require `meroxa login` for any command.
-			There are command that don't require client such as `meroxa env`, and we wouldn't like to throw an error,
-			just because we can't emit events.
-	*/
-	if err != nil {
-		return "", "", nil
-	}
-
-	// fetch actor account.
-	actor = global.Config.GetString("MEROXA_ACTOR")
-	actorUUID = global.Config.GetString("MEROXA_ACTOR_UUID")
-
-	if actor == "" || actorUUID == "" {
-		// call api to fetch
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second) // nolint:gomnd
-		defer cancel()
-
-		m, err := global.NewClient()
-
-		if err != nil {
-			return "", "", fmt.Errorf("meroxa: could not create Meroxa client: %v", err)
-		}
-
-		account, err := m.GetUser(ctx)
-
-		if err != nil {
-			return "", "", fmt.Errorf("meroxa: could not fetch Meroxa user: %v", err)
-		}
-
-		actor = account.Email
-		actorUUID = account.UUID
-
-		global.Config.Set("MEROXA_ACTOR", actor)
-		global.Config.Set("MEROXA_ACTOR_UUID", actorUUID)
-
-		err = global.Config.WriteConfig()
-
-		if err != nil {
-			if _, ok := err.(viper.ConfigFileNotFoundError); ok {
-				err = global.Config.SafeWriteConfig()
-			}
-			if err != nil {
-				return "", "", fmt.Errorf("meroxa: could not write config file: %v", err)
-			}
-		}
-	}
-
-	return actor, actorUUID, nil
-}
-
 // This runs for all commands.
 func buildCommandWithEvent(cmd *cobra.Command, c Command) {
 	oldRunE := cmd.RunE
@@ -397,40 +341,7 @@ func buildCommandWithEvent(cmd *cobra.Command, c Command) {
 				return err
 			}
 
-			// This will be empty on logout, maybe do it before?
-			actor, actorUUID, err := getCLIUserInfo()
-
-			if err != nil {
-				return err
-			}
-
-			// build our event
-			event := cased.AuditEvent{
-				"timestamp":  time.Now().UTC(),
-				"user_agent": fmt.Sprintf("meroxa/%s %s/%s", global.Version, runtime.GOOS, runtime.GOARCH),
-			}
-
-			if actor != "" {
-				event["actor"] = actor
-			}
-
-			if actorUUID != "" {
-				event["actor_uuid"] = actorUUID
-			}
-
-			var action string
-
-			// TODO: Implement something that could look up all the way up until meroxa (meroxa create resources...)
-			// something like it determines how many levels since root and then until current cmd
-			if cmd.HasParent() {
-				if cmd.Parent().HasParent() {
-					action = fmt.Sprintf("%s.%s.%s", cmd.Parent().Parent().Use, cmd.Parent().Use, cmd.Use)
-				} else {
-					action = fmt.Sprintf("%s.%s", cmd.Parent().Use, cmd.Use)
-				}
-			} else {
-				action = cmd.Use
-			}
+			event := global.BuildEvent(cmd)
 
 			if cmd.Use != cmd.CalledAs() {
 				event["command.alias"] = cmd.CalledAs()
@@ -439,9 +350,6 @@ func buildCommandWithEvent(cmd *cobra.Command, c Command) {
 			if len(args) > 0 {
 				event["command.args"] = args
 			}
-
-			event["action"] = action
-			event["use"] = action
 
 			if err != nil {
 				event["error"] = err
